glyph_forge/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "glyph_forge"
version = "0.1.0"
description = "Eidosian Glyph Forge - Image to ASCII/ANSI Converter."
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [
    { name = "Lloyd Handyside", email = "ace1928@gmail.com" },
    { name = "Eidos", email = "syntheticeidos@gmail.com" },
]
keywords = ["ascii-art", "ansi", "image-processing", "pillow"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pillow>=9.0.0",
    "numpy>=1.26.0,<2.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
    "pyfiglet>=0.8.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-benchmark>=5.2.0",
    "pytest-cov>=6.0.0",
    "pytest-html>=4.2.0",
    "pytest-mock>=3.15.0",
    "pytest-profiling>=1.8.0",
]
streaming = [
    "opencv-python>=4.8.0,<4.13",
    "yt-dlp>=2024.12.0",
    "mss>=9.0.0",
]
browser = [
    "playwright>=1.40.0",
]
tui = [
    "textual>=0.50.0",
]
audio = [
    "pygame>=2.5.0",
    "simpleaudio>=1.0.4",
]
assets = [
    "requests>=2.31.0",
]
clipboard = [
    "pyperclip>=1.9.0",
]
all = [
    "pytest>=8.0.0",
    "pytest-benchmark>=5.2.0",
    "pytest-cov>=6.0.0",
    "pytest-html>=4.2.0",
    "pytest-mock>=3.15.0",
    "pytest-profiling>=1.8.0",
    "opencv-python>=4.8.0,<4.13",
    "yt-dlp>=2024.12.0",
    "mss>=9.0.0",
    "playwright>=1.40.0",
    "textual>=0.50.0",
    "pygame>=2.5.0",
    "simpleaudio>=1.0.4",
    "requests>=2.31.0",
    "pyperclip>=1.9.0",
]

[project.scripts]
glyph-forge = "glyph_forge.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/glyph_forge"]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = ["tests"]
markers = [
    "integration: mark a test as an integration test.",
]
